+++ jabls/src/main/java/module-info.java
@@ -1,7 +1,8 @@
 module org.jabref.jabls {
     exports org.jabref.languageserver;
-    opens org.jabref.languageserver to lsp4j, lsp4j.jsonrpc;
+    opens org.jabref.languageserver to lsp4j, lsp4j.jsonrpc, com.google.gson;
     exports org.jabref.languageserver.manager;
+    exports org.jabref.languageserver.util;
 
     requires org.jabref.jablib;
 
@@ -10,4 +11,5 @@ module org.jabref.jabls {
     requires lsp4j;
     requires lsp4j.jsonrpc;
     requires lsp4j.websocket;
+    requires com.google.gson;
 }